Refactor: 일요일과 토요일 날짜의 문자열 포멧을 만들어 줄 유틸함수 정의 및 사용
Changes to be committed:
modified:   src/hooks/page/useExerciseRoutine.ts
new file:   src/utils/date.ts
modified:   src/utils/index.ts
