+++ src/hooks/page/useExerciseRoutine.ts
@@ -1,31 +1,22 @@
 import { useFetchExerciseHistory, useFetchExerciseSuggestion } from '@/hooks/api';
+import { getFirstAndLastDateStringFromCurrentWeek } from '@/utils';
 
 const useExerciseRoutine = () => {
-  const today = new Date();
-  const currentYear = today.getFullYear();
-  const currentMonth = today.getMonth();
-  const currentDate = today.getDate();
-  const currentDay = today.getDay();
-
-  const sunday = new Date(currentYear, currentMonth, currentDate - currentDay);
-  const saturday = new Date(currentYear, currentMonth, currentDate + (6 - currentDay));
-
-  const from = `${sunday.getFullYear()}-${sunday.getMonth() + 1}-${sunday.getDate()}`;
-  const to = `${saturday.getFullYear()}-${saturday.getMonth() + 1}-${saturday.getDate()}`;
+  const { sundayDateString, saturdayDateString } = getFirstAndLastDateStringFromCurrentWeek();
 
   const {
     suggestion,
     error: suggestionError,
     isError: isSuggestionError,
     isLoading: isLoadingSuggestion,
-  } = useFetchExerciseSuggestion(from, to);
+  } = useFetchExerciseSuggestion(sundayDateString, saturdayDateString);
 
   const {
     exerciseHistory,
     error: exerciseHistoryError,
     isError: isExerciseHistoryError,
     isLoading: isLoadingExerciseHistory,
-  } = useFetchExerciseHistory(from, to);
+  } = useFetchExerciseHistory(sundayDateString, saturdayDateString);
 
   const error = suggestionError || exerciseHistoryError;
   const isError = isSuggestionError || isExerciseHistoryError;
